Day5 cleanup
Cleaned up the code!
