+++ Day 5/day5_funcs.h
@@ -1,0 +1,15 @@
+#include <vector>
+#include <string>
+#include <regex>
+
+using namespace std;
+
+int checkForDoubleLetters(string inputString);
+
+int countVowels(string inputString);
+
+int checkforDoublePairs(string inputString);
+
+int checkForXyXFlag(string inputString);
+
+int checkForNaughtyPairs(string inputString, vector<regex> badStrings);
